+++ assets/javascripts/kitten/components/structure/expand/dropdown-menu/stories.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { action } from '@storybook/addon-actions'
 import { DropdownMenu } from './index'
-import { COLORS, EllipsisIcon } from 'kitten'
+import { COLORS, EllipsisIcon, AvatarWithTextAndBadge, Text, ArrowIcon } from 'kitten'
 import styled from 'styled-components'
 import { DocsPage } from 'storybook/docs-page'
 
@@ -106,3 +106,43 @@ export const Default = ({ menuPosition, ...args }) => {
     </StyledWrapper>
   )
 }
+
+const StyledAvatarWrapper = styled.div`
+  display: block;
+  position: relative;
+`
+
+export const WithAvatar = args => (
+  <StyledAvatarWrapper>
+    <DropdownMenu
+      {...args}
+      menuPosition="right"
+      button={(open) => (
+        <AvatarWithTextAndBadge>
+          <AvatarWithTextAndBadge.Image
+            alt=""
+            src="/kitten-0.jpg"
+          />
+
+          <AvatarWithTextAndBadge.Text>
+            <div>
+              <Text weight="regular" className="k-u-block k-u-link k-u-link-primary1">Kitten Kitty</Text>
+              <Text className="k-u-block">Cat City</Text>
+            </div>
+          </AvatarWithTextAndBadge.Text>
+          <ArrowIcon
+            className="k-u-flex-shrink-none"
+            direction={open ? 'top' : 'bottom'}
+          />
+        </AvatarWithTextAndBadge>
+      )}
+    >
+      <DropdownMenu.Link href="">A link</DropdownMenu.Link>
+      <DropdownMenu.Button>Click this button</DropdownMenu.Button>
+      <DropdownMenu.Link href="">
+        Another last very very very very very very very very very very long
+        link
+      </DropdownMenu.Link>
+    </DropdownMenu>
+  </StyledAvatarWrapper>
+)