+++ CHANGELOG.md
@@ -9,7 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: `HeaderNav`: Add `a11yMobileOpen` and `a11yMobileClose` props to `SearchInput`.
 - Feature: `Carousel`: More accessible handling of prev/next click.
 - Fix: `DropdownMenu`: Fix button argument.
-- Fix: Utilities: Add `k-utilities-flexbox-element` utility classes.
+- Feature: Utilities: Add `k-utilities-flexbox-element` utility classes.
+- Fix: `DropdownMenu`: Structure fixes.
 
 ## [8.3.1] - 2022-02-10